- upload --resume switch implemented (requires bookkeeping). - switched to using archive_id as item key in the bookkeeping db. - added updatedb command for people that are upgrading to change their db. - updated readme to reflect changes. - some minor clean-up of the files.
Fix for updatedb.
It didn't work as it should...

Fix for Python <2.7

mmapping file now in parts, instead of trying to mmap it completely.

Fixed upload of large files.
Mmap has a 2 GB limit; now mapping the relevant part of the files (either the part to be checked or the part to be uploaded) instead of attempting to map the complete file in one go.

Hope to have improved reaction times on the connection (see issue #71)

Added resumption of downloads (untested).

Download resumption stage 1.
Works mostly (hash check currently fails and needs work).

Implemented download resumption.

If the flag --resume is given, glacier-cmd will check existing data in <out_file> and if it matches, continue the download from where it was.

fix

Implemented parallel upload sessions.

Added command line --sessions to give the number of upload sessions to use.

It is far from perfect: at the moment issues with upload processes dying (amazon rejects our signature, as if the key is invalid?!) and then the whole thing just hangs and needs to be killed manually. This needs a workaround.

Updated documentation to include the new options and commands.