- upload --resume switch implemented (requires bookkeeping). - switched to using archive_id as item key in the bookkeeping db. - added updatedb command for people that are upgrading to change their db. - updated readme to reflect changes. - some minor clean-up of the files.
Fix for updatedb.
It didn't work as it should...

Fix for Python <2.7

mmapping file now in parts, instead of trying to mmap it completely.

Fixed upload of large files.
Mmap has a 2 GB limit; now mapping the relevant part of the files (either the part to be checked or the part to be uploaded) instead of attempting to map the complete file in one go.

Hope to have improved reaction times on the connection (see issue #71)

Implemented automatic upload resumption.
Implemented speed fix (see issue #71).
Fixed up bookkeeping of uploads.
Changed mmapping of files to allow for big files.
Added --updatedb switch to update the db from the old style (file name as key) to new style (archive_id as key).

- upload --resume switch implemented (requires bookkeeping).
- switched to using archive_id as item key in the bookkeeping db.
- added updatedb command for people that are upgrading to change their db.
- updated readme to reflect changes.
- some minor clean-up of the files.

Fix for updatedb.
It didn't work as it should...

mmapping file now in parts, instead of trying to mmap it completely.

merge

- upload --resume switch implemented (requires bookkeeping).
- switched to using archive_id as item key in the bookkeeping db.
- added updatedb command for people that are upgrading to change their db.
- updated readme to reflect changes.
- some minor clean-up of the files.

Fix for updatedb.
It didn't work as it should...

Fix for Python <2.7

mmapping file now in parts, instead of trying to mmap it completely.

Fixed upload of large files.
Mmap has a 2 GB limit; now mapping the relevant part of the files (either the part to be checked or the part to be uploaded) instead of attempting to map the complete file in one go.

Hope to have improved reaction times on the connection (see issue #71)

merge