Finally bookkeeping=False or bookkeeping=True is handled correctly!
The problem: `default("bookkeeping")` gives the result 'True' or 'False' as string value, not the boolean True or boolean False. So we must use string comparison, not boolean comparison. Tricky!

Also added check in sdb_connect that we actually have a domain if bookkeeping is switched on; if not it throws an exception telling the user to set a domain or to switch off bookkeeping alltogether.