Implemented parallel upload sessions.
Added command line --sessions to give the number of upload sessions to use.

It is far from perfect: at the moment issues with upload processes dying (amazon rejects our signature, as if the key is invalid?!) and then the whole thing just hangs and needs to be killed manually. This needs a workaround.